requirements.txt
fastapi==0.119.1
uvicorn==0.38.0
streamlit==1.50.0
pandas==2.3.3
matplotlib==3.10.7
pytesseract==0.3.13
opencv-python==4.12.0.88
python-multipart==0.0.20
requests==2.32.5
torch==2.9.0
transformers==4.57.1
pillow==11.3.0
